+++ thoth-app/src/component/work.rs
@@ -306,7 +306,6 @@ impl Component for WorkComponent {
                     self.work.page_interval = None;
                 }
                 if self.work.work_status != WorkStatus::WithdrawnFromSale
-                    && self.work.work_status != WorkStatus::OutOfPrint
                 {
                     self.work.withdrawn_date = None;
                 }
@@ -578,9 +577,8 @@ impl Component for WorkComponent {
                 // Grey out chapter-specific or "book"-specific fields
                 // based on currently selected work type.
                 let is_chapter = self.work.work_type == WorkType::BookChapter;
-                let is_not_withdrawn_or_out_of_print = self.work.work_status
-                    != WorkStatus::WithdrawnFromSale
-                    && self.work.work_status != WorkStatus::OutOfPrint;
+                let is_not_withdrawn = self.work.work_status
+                    != WorkStatus::WithdrawnFromSale;
                 html! {
                     <>
                         <nav class="level">
@@ -661,7 +659,7 @@ impl Component for WorkComponent {
                                 value={ self.work.withdrawn_date.clone() }
                                 oninput={ ctx.link().callback(|e: InputEvent| Msg::ChangeWithdrawnDate(e.to_value())) }
                                 required = true
-                                deactivated={ is_not_withdrawn_or_out_of_print }
+                                deactivated={ is_not_withdrawn }
                                 />
                             <FormTextInput
                                 label = "Place of Publication"