+++ thoth-export-server/src/xml/onix3_google_books.rs
@@ -504,19 +504,12 @@ impl XmlElement<Onix3GoogleBooks> for WorkStatus {
 
     fn value(&self) -> &'static str {
         match self {
-            WorkStatus::UNSPECIFIED => "00",
             WorkStatus::CANCELLED => "01",
             WorkStatus::FORTHCOMING => "02",
             WorkStatus::POSTPONED_INDEFINITELY => "03",
             WorkStatus::ACTIVE => "04",
-            WorkStatus::NO_LONGER_OUR_PRODUCT => "05",
-            WorkStatus::OUT_OF_STOCK_INDEFINITELY => "06",
-            WorkStatus::OUT_OF_PRINT => "07",
-            WorkStatus::INACTIVE => "08",
-            WorkStatus::UNKNOWN => "09",
-            WorkStatus::REMAINDERED => "10",
-            // 15 Recalled is not supported by Google Books
-            WorkStatus::WITHDRAWN_FROM_SALE | WorkStatus::RECALLED => "11",
+            WorkStatus::SUPERSEDED => "08",
+            WorkStatus::WITHDRAWN_FROM_SALE => "11",
             WorkStatus::Other(_) => unreachable!(),
         }
     }